plugins/inference/mir-templates/pyproject.toml
# (C) Copyright 2025- ECMWF.
#
# This software is licensed under the terms of the Apache Licence Version 2.0
# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
#
# In applying this licence, ECMWF does not waive the privileges and immunities
# granted to it by virtue of its status as an intergovernmental organisation
# nor does it submit to any jurisdiction.

[build-system]
requires = [ "setuptools>=60", "setuptools-scm>=8" ]

[project]
name = "anemoi-inference-grib-templates-mir"

description = "A grib templates provider with mir"
readme = "README.md"

keywords = [
  "ai",
  "inference",
  "output",
  "tools",
]

license = { file = "LICENSE" }
authors = [
  { name = "European Centre for Medium-Range Weather Forecasts (ECMWF)", email = "software.support@ecmwf.int" },
]
requires-python = ">=3.10"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]

dynamic = [ "version" ]

dependencies = [
  "anemoi-inference>=0.5.1",
  "earthkit-data",
  "mir-python>=1.27.1.dev20250627",
]
optional-dependencies.tests = [ "anemoi-inference[tests]>=0.5.2", "pytest", "pytest-xdist" ]

urls.Homepage = "https://github.com/ecmwf/anemoi-plugins"

entry-points."anemoi.inference.grib.templates".mir = "anemoi_inference_grib_templates_mir.mir_templates:MirTemplatesProvider"

[tool.setuptools.package-data]
"anemoi_inference_grib_templates_mir" = [ "*.yaml" ]

[tool.setuptools_scm]

root = ".."
version_scheme = "post-release"
local_scheme = "no-local-version"
git_describe_command = "git describe --dirty --tags --long --match 'inference-mir-*[0-9]*'"
tag_regex = "^inference-mir-(?P<version>[vV]?[0-9]+[^-]*)"
version_file = "src/anemoi_inference_grib_templates_mir/_version.py"
fallback_version = "0.0.0"

[tool.pytest.ini_options]
testpaths = "tests"
addopts = [
  "--numprocesses=8",
  "--strict-config",
]

[tool.uv.pip]
prerelease = "allow"
